pyproject.toml
[tool.poetry]
name = "horseclients"
version = "0.1.0"
description = ""
authors = ["artyom <artyom27000@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyaudio = "^0.2.14"
scipy = "^1.11.4"
grpcio-tools = "^1.48.2"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
